refactor: generalize the DataTable
